Move FV0 MCLabel to DataFormats
