+++ DataFormats/Detectors/FIT/FV0/CMakeLists.txt
@@ -21,6 +21,7 @@ o2_add_library(DataFormatsFV0
 o2_target_root_dictionary(DataFormatsFV0
         HEADERS include/DataFormatsFV0/Hit.h
                 include/DataFormatsFV0/BCData.h
+                include/DataFormatsFV0/MCLabel.h
                 include/DataFormatsFV0/ChannelData.h
                 include/DataFormatsFV0/RawEventData.h
                 include/DataFormatsFV0/LookUpTable.h)
