+++ Detectors/FIT/FV0/simulation/src/FV0SimulationLinkDef.h
@@ -17,7 +17,6 @@
 #pragma link C++ class o2::fv0::Detector + ;
 #pragma link C++ class o2::base::DetImpl < o2::fv0::Detector> + ;
 #pragma link C++ class o2::fv0::Digitizer + ;
-#pragma link C++ class o2::fv0::MCLabel + ;
 #pragma link C++ class o2::dataformats::MCTruthContainer < o2::fv0::MCLabel> + ;
 #pragma link C++ class o2::fv0::DigitizationConstant + ;
 #pragma link C++ class o2::fv0::FV0DigParam + ;
